Before sending to site

+++ php/parser_dayly.php
@@ -1,5 +1,5 @@
 <?php
-    //require_once('parser_start.php');
+    
     header('Content-type: text/html; charset=utf-8');
     require_once('classes/simple_html_dom.php');
     require_once('classes/Parse.php');
@@ -9,11 +9,13 @@
     $descrMax = 799; // Максимальное количество символов в описании
     $period = 'Day';  // 'Day' or 'All'
 
-    mysqli_query($mysqli, "SET FOREIGN_KEY_CHECKS = 0");
-    mysqli_query($mysqli,"DELETE FROM `excursion` WHERE `date` BETWEEN CURRENT_DATE() - INTERVAL 1 DAY AND CURRENT_DATE() - INTERVAL 1 DAY");
-    
+    if ($period == 'Day'){
+        mysqli_query($mysqli, "SET FOREIGN_KEY_CHECKS = 0");
+        mysqli_query($mysqli,"DELETE FROM `excursion` WHERE `date` BETWEEN CURRENT_DATE() - INTERVAL 1 DAY AND CURRENT_DATE() - INTERVAL 1 DAY");
+    }
+
     Parse::parseMS($period);
-    Parse::parseMW($period);
+    Parse::parseMW();
     Parse::parseTM($period);
     Parse::parseMV($period);
     Parse::parseMH($period);
